extract the cord area vs mJOA score Linear Mixed-Effects Models analysis into a separate script
